Use LanguageClient and deoplete. Organize leader
LanguageClient-Neovim provides a whole lot of functionality that I wish
to start using. Almost all the languages I use except for go will use
the languageclient to some extent. Deoplete seemed to work better for
this, and since it has not crashed in go files recently, we will see
where this goes. If it works, I will remove the commented lines for
nvim-completion-manager. A function for setting up the mappings for
languageclient was created to keep things DRY

Leader functions have been moved to their own file in leader.vim